+++ R/calculate_geocbi.R
@@ -32,7 +32,7 @@
 #'
 #' @param s1_fcov_post_fire_leaf_fall A numeric grade from 0.0 to 3.0.
 #'   Post-fire leaf fall cover.
-#'   Includes leaves that have fallen after the burn but not unburned patches.
+#'   Includes leaves that have fallen after the burn but not unburnt patches.
 #' @param s1_area_unburnt A numeric grade from 0.0 to 3.0.
 #'   Unburnt area.
 #'   The approximate percentage of area not burned.
@@ -100,7 +100,8 @@
 #'   s1_fcov_post_fire_leaf_fall = 1,
 #'   s2_area_unburnt = 2.5,
 #'   s2_fcov = 0.5,
-#'   verbose = TRUE)
+#'   verbose = TRUE
+#' )
 #'
 #' # With complete variables, all set to 1
 #' calculate_geocbi(
@@ -121,12 +122,12 @@
 #'   s5_no_trees_over_20cm_fallen = 1,
 #'   s5_resprouting_on_live_trees = 1,
 #'   s5_char_height = 1,
-#'   s1_fcov = 1,
 #'   s2_fcov = 1,
 #'   s3_fcov = 1,
 #'   s4_fcov = 1,
 #'   s5_fcov = 1,
-#'   verbose = TRUE)
+#'   verbose = TRUE
+#' )
 calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
                              s1_area_unburnt = NA_real_,
                              s1_duff = NA_real_,
@@ -163,7 +164,7 @@ calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
     purrr::discard(is.na) %>%
     mean()
   s1_score <- s1_cbi * s1_fcov
-  if (verbose == T) {
+  if (verbose == TRUE) {
     message(
       glue::glue("Stratum 1: CBI {s1_cbi} * FCOV {s1_fcov} = Score {s1_score}")
     )
@@ -181,7 +182,7 @@ calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
     purrr::discard(is.na) %>%
     mean()
   s2_score <- s2_cbi * s2_fcov
-  if (verbose == T) {
+  if (verbose == TRUE) {
     message(
       glue::glue("Stratum 2: CBI {s2_cbi} * FCOV {s2_fcov} = Score {s2_score}")
     )
@@ -196,7 +197,7 @@ calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
     purrr::discard(is.na) %>%
     mean()
   s3_score <- s3_cbi * s3_fcov
-  if (verbose == T) {
+  if (verbose == TRUE) {
     message(
       glue::glue("Stratum 3: CBI {s3_cbi} * FCOV {s3_fcov} = Score {s3_score}")
     )
@@ -212,7 +213,7 @@ calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
     purrr::discard(is.na) %>%
     mean()
   s4_score <- s4_cbi * s4_fcov
-  if (verbose == T) {
+  if (verbose == TRUE) {
     message(
       glue::glue("Stratum 4: CBI {s4_cbi} * FCOV {s4_fcov} = Score {s4_score}")
     )
@@ -229,7 +230,7 @@ calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
     purrr::discard(is.na) %>%
     mean()
   s5_score <- s5_cbi * s5_fcov
-  if (verbose == T) {
+  if (verbose == TRUE) {
     message(
       glue::glue("Stratum 5: CBI {s5_cbi} * FCOV {s5_fcov} = Score {s5_score}")
     )
@@ -239,12 +240,13 @@ calculate_geocbi <- function(s1_fcov_post_fire_leaf_fall = NA_real_,
   score_sum <- sum(s1_score, s2_score, s3_score, s4_score, s5_score, na.rm = T)
   fcov_sum <- sum(s1_fcov, s2_fcov, s3_fcov, s4_fcov, s5_fcov)
   geocbi <- score_sum / fcov_sum
-  if (verbose == T)
+  if (verbose == TRUE) {
     message(
       glue::glue(
         "GeoCBI: {geocbi} = Score sum {score_sum} / FCOV sums {fcov_sum}"
-        )
+      )
     )
+  }
   geocbi
 }