+++ NEWS.md
@@ -1,7 +1,9 @@
 # rGeoCBI 0.0.2
+
 * Some improvements to form "Burn Grading 0.1" identified and in preparation
   as "Burn Grading 0.2". Pending review and QA.
 * Add workhorse function `calculate_geocbi`. Example see vignette "analysis".
+* Move helper `map_burngrading` from vignette to own function.
 
 # rGeoCBI 0.0.1